feat: relocate currency detection logic to invoice_models.py for better organization and maintainability
